Tasks
Task 1 - To complete this challenge, you must print Hello Solidity. Use str as a variable to store value.
Task 2 - 
Create functions
Task-

To complete this challenge, you have to create a state variable and initialised it with 10 and then you have to create a function returnStateVariable( ) which returns the value of the created state variable .

In the same way you have to create a local variable and initialised it with 20 and then you have to create a function returnLocalVariable( ) which returns the value of the created local variable.